try to fix massive memory used if writing large BAM files
this involves writing them pre-sorted as otherwise Picard accumulates
records in memory and creates very large temporary files. There
are two aspects to that: when iterating using eachPair we have to
try very hard to eliminate scenarios where we iterate the pairs
out of order. Previously that could occur in various situations.
Then when writing, we still have to buffer reads in some circumstances,
so the OrderedPairWriter now supports this.